# backend/artifacts/openapi-v6.0.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ClearPath Nexus Engine",
    "version": "6.0.0"
  },
  "paths": {
...
    "/api/v1/port/berths": {
      "get": {
        "tags": [
          "port"
        ],
        "summary": "List Berths",
        "operationId": "list_berths_api_v1_port_berths_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PortBerthResponse"
                  },
                  "type": "array",
                  "title": "Response List Berths Api V1 Port Berths Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/port/sync-status": {
      "get": {
        "tags": [
          "port"
        ],
        "summary": "Port Sync Status",
        "operationId": "port_sync_status_api_v1_port_sync_status_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "port_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "JNPT_MUMBAI",
              "title": "Port Id"
            }
          },
          {
            "name": "vessel_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "MAERSK_X26",
              "title": "Vessel Id"
            }
          },
          {
            "name": "train_arrival_hours",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "default": 24.0,
              "title": "Train Arrival Hours"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortSyncStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/provenance/routes/{route_id}": {
      "get": {
        "tags": [
          "provenance"
        ],
        "summary": "Get Route Evidence",
        "description": "Return the stored decision evidence. This endpoint never refetches providers.",
        "operationId": "get_route_evidence_api_v1_provenance_routes__route_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "route_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Route Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteEvidenceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/provenance/records/{record_id}": {
      "get": {
        "tags": [
          "provenance"
        ],
        "summary": "Get Record Neighborhood",
        "operationId": "get_record_neighborhood_api_v1_provenance_records__record_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Record Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordNeighborhoodResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/provenance/sources": {
      "get": {
        "tags": [
          "provenance"
        ],
        "summary": "List Sources",
        "operationId": "list_sources_api_v1_provenance_sources_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceListResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EvidenceComponent": {
        "properties": {
          "role": {
            "type": "string",
            "title": "Role"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "record_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Record Ids"
          },
          "explanation": {
            "type": "string",
            "title": "Explanation"
          }
        },
        "type": "object",
        "required": [
          "role",
          "label",
          "status",
          "record_ids",
          "explanation"
        ],
        "title": "EvidenceComponent"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LineageEdgeResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "parent_record_id": {
            "type": "string",
            "format": "uuid",
            "title": "Parent Record Id"
          },
          "child_record_id": {
            "type": "string",
            "format": "uuid",
            "title": "Child Record Id"
          },
          "relationship": {
            "type": "string",
            "title": "Relationship"
          }
        },
        "type": "object",
        "required": [
          "id",
          "parent_record_id",
          "child_record_id",
          "relationship"
        ],
        "title": "LineageEdgeResponse"
      },
...
      "PortBerthResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "port_name": {
            "type": "string",
            "title": "Port Name"
          },
          "berth_identifier": {
            "type": "string",
            "title": "Berth Identifier"
          },
          "vessel_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vessel Name"
          },
          "window_start": {
            "type": "string",
            "format": "date-time",
            "title": "Window Start"
          },
          "window_end": {
            "type": "string",
            "format": "date-time",
            "title": "Window End"
          }
        },
        "type": "object",
        "required": [
          "id",
          "port_name",
          "berth_identifier",
          "vessel_name",
          "window_start",
          "window_end"
        ],
        "title": "PortBerthResponse"
      },
...
      "PortSyncStatus": {
        "properties": {
          "aligned": {
            "type": "boolean",
            "title": "Aligned"
          },
          "berth_id": {
            "type": "string",
            "title": "Berth Id"
          },
          "vessel_status": {
            "type": "string",
            "title": "Vessel Status"
          },
          "loading_window_start": {
            "type": "string",
            "format": "date-time",
            "title": "Loading Window Start"
          },
          "loading_window_end": {
            "type": "string",
            "format": "date-time",
            "title": "Loading Window End"
          },
          "sync_score": {
            "type": "number",
            "title": "Sync Score"
          },
          "warning": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Warning"
          }
        },
        "type": "object",
        "required": [
          "aligned",
          "berth_id",
          "vessel_status",
          "loading_window_start",
          "loading_window_end",
          "sync_score"
        ],
        "title": "PortSyncStatus"
      },
...
      "ProvenanceRecordResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "entity_type": {
            "type": "string",
            "title": "Entity Type"
          },
          "entity_key": {
            "type": "string",
            "title": "Entity Key"
          },
          "decision_input_role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision Input Role"
          },
          "canonical_source_type": {
            "type": "string",
            "title": "Canonical Source Type"
          },
          "raw_source_state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Source State"
          },
          "source": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SourceBrief"
              },
              {
                "type": "null"
              }
            ]
          },
          "observed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observed At"
          },
          "fetched_at": {
            "type": "string",
            "format": "date-time",
            "title": "Fetched At"
          },
          "valid_until": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid Until"
          },
          "freshness": {
            "additionalProperties": true,
            "type": "object",
            "title": "Freshness"
          },
          "cache_hit": {
            "type": "boolean",
            "title": "Cache Hit"
          },
          "used_in_decision": {
            "type": "boolean",
            "title": "Used In Decision"
          },
          "excluded_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Excluded Reason"
          },
          "quality": {
            "additionalProperties": true,
            "type": "object",
            "title": "Quality"
          },
          "transform": {
            "additionalProperties": true,
            "type": "object",
            "title": "Transform"
          },
          "decision_impact": {
            "additionalProperties": true,
            "type": "object",
            "title": "Decision Impact"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "checksum": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checksum"
          },
          "value_summary": {
            "additionalProperties": true,
            "type": "object",
            "title": "Value Summary"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "id",
          "entity_type",
          "entity_key",
          "decision_input_role",
          "canonical_source_type",
          "raw_source_state",
          "source",
          "observed_at",
          "fetched_at",
          "valid_until",
          "freshness",
          "cache_hit",
          "used_in_decision",
          "excluded_reason",
          "quality",
          "transform",
          "decision_impact",
          "request_id",
          "checksum",
          "value_summary",
          "metadata"
        ],
        "title": "ProvenanceRecordResponse"
      },
      "ProvenanceSummary": {
        "properties": {
          "decision_record_id": {
            "type": "string",
            "format": "uuid",
            "title": "Decision Record Id"
          },
          "traceability": {
            "$ref": "#/components/schemas/TraceabilityCounts"
          },
          "freshness": {
            "$ref": "#/components/schemas/FreshnessCounts"
          },
          "source_modes": {
            "$ref": "#/components/schemas/SourceModeCounts"
          },
          "decision_use": {
            "$ref": "#/components/schemas/DecisionUseCounts"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "decision_record_id",
          "traceability",
          "freshness",
          "source_modes",
          "decision_use"
        ],
        "title": "ProvenanceSummary"
      },
...
      "RecordNeighborhoodResponse": {
        "properties": {
          "record": {
            "$ref": "#/components/schemas/ProvenanceRecordResponse"
          },
          "parents": {
            "items": {
              "$ref": "#/components/schemas/ProvenanceRecordResponse"
            },
            "type": "array",
            "title": "Parents"
          },
          "children": {
            "items": {
              "$ref": "#/components/schemas/ProvenanceRecordResponse"
            },
            "type": "array",
            "title": "Children"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/LineageEdgeResponse"
            },
            "type": "array",
            "title": "Edges"
          }
        },
        "type": "object",
        "required": [
          "record",
          "parents",
          "children",
          "edges"
        ],
        "title": "RecordNeighborhoodResponse"
      },
...
      "RouteEvidenceResponse": {
        "properties": {
          "route": {
            "additionalProperties": true,
            "type": "object",
            "title": "Route"
          },
          "snapshot": {
            "additionalProperties": true,
            "type": "object",
            "title": "Snapshot"
          },
          "traceability": {
            "$ref": "#/components/schemas/ProvenanceSummary"
          },
          "components": {
            "items": {
              "$ref": "#/components/schemas/EvidenceComponent"
            },
            "type": "array",
            "title": "Components"
          },
          "records": {
            "items": {
              "$ref": "#/components/schemas/ProvenanceRecordResponse"
            },
            "type": "array",
            "title": "Records"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/LineageEdgeResponse"
            },
            "type": "array",
            "title": "Edges"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "route",
          "snapshot",
          "traceability",
          "components",
          "records",
          "edges",
          "warnings"
        ],
        "title": "RouteEvidenceResponse"
      },
...
      "SourceBrief": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "key": {
            "type": "string",
            "title": "Key"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "authority_level": {
            "type": "string",
            "title": "Authority Level"
          },
          "official": {
            "type": "boolean",
            "title": "Official"
          },
          "free_for_mvp": {
            "type": "boolean",
            "title": "Free For Mvp"
          },
          "requires_key": {
            "type": "boolean",
            "title": "Requires Key"
          },
          "reference_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference Url"
          },
          "license_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "License Name"
          },
          "attribution_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attribution Text"
          },
          "limitations": {
            "additionalProperties": true,
            "type": "object",
            "title": "Limitations"
          },
          "freshness_policy": {
            "additionalProperties": true,
            "type": "object",
            "title": "Freshness Policy"
          },
          "current_status": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "key",
          "label",
          "category",
          "authority_level",
          "official",
          "free_for_mvp",
          "requires_key"
        ],
        "title": "SourceBrief"
      },
      "SourceListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SourceBrief"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "SourceListResponse"
      },
...
      "TraceabilityCounts": {
        "properties": {
          "traced": {
            "type": "integer",
            "title": "Traced"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "coverage_pct": {
            "type": "number",
            "title": "Coverage Pct"
          }
        },
        "type": "object",
        "required": [
          "traced",
          "total",
          "coverage_pct"
        ],
        "title": "TraceabilityCounts"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}